feat: merge link and base link
